+++ assets/javascripts/kitten/helpers/utils/parser.js
@@ -1,0 +1,3 @@
+import HtmlToReact from 'html-to-react'
+
+export const parseHtml = value => new HtmlToReact.Parser().parse(value)